Mise à jour de WordPress, plug-ins, thèmes, traductions. Désactivation du plug-in « WP My Social Networks », qui affiche une erreur.

+++ wp-content/plugins/nextgen-gallery/products/photocrati_nextgen/modules/ngglegacy/admin/manage.php
@@ -17,7 +17,7 @@ class nggManageGallery {
 		// GET variables
 		if( isset($_GET['gid']) ) {
 			$this->gid  = (int) $_GET['gid'];
-			$this->gallery = C_Gallery_Mapper::get_instance()->find($this->gid);
+			$this->gallery = C_Gallery_Mapper::get_instance()->find($this->gid, TRUE);
 		}
 		if( isset($_GET['pid']) )
 			$this->pid  = (int) $_GET['pid'];
@@ -31,14 +31,11 @@ class nggManageGallery {
             }
         }
         // Should be only called via manage galleries overview
-		if ( isset($_POST['page']) && $_POST['page'] == 'manage-galleries' )
+		if ( isset($_POST['nggpage']) && $_POST['nggpage'] == 'manage-galleries' )
 			$this->post_processor_galleries();
 		// Should be only called via a edit single gallery page
-		if ( isset($_POST['page']) && $_POST['page'] == 'manage-images' )
+		if ( isset($_POST['nggpage']) && $_POST['nggpage'] == 'manage-images' )
 			$this->post_processor_images();
-		// Should be called via a publish dialog
-		if ( isset($_POST['page']) && $_POST['page'] == 'publish-post' )
-			$this->publish_post();
 
 		//Look for other POST process
 		if ( !empty($_POST) || !empty($_GET) )
@@ -194,7 +191,7 @@ class nggManageGallery {
 
 		$output = array();
 
-		$output[] = "<a href='{$image_url}' class='shutter' title='{$filename}'>";
+		$output[] = "<a href='{$image_url}' class='thickbox' title='{$filename}'>";
 		$output[] = "<img class='thumb' src='{$thumb_url}' id='thumb{$picture->pid}'/>";
 		$output[] = "</a>";
 
@@ -262,7 +259,6 @@ class nggManageGallery {
 			'meta'			=>	array(&$this, 'render_meta_action_link'),
 			'custom_thumb'	=>	array(&$this, 'render_custom_thumb_action_link'),
 			'rotate'		=>	array(&$this, 'render_rotate_action_link'),
-			'publish'		=>	array(&$this, 'render_publish_action_link'),
 			'recover'		=>	array(&$this, 'render_recover_action_link'),
 			'delete'		=>	array(&$this, 'render_delete_action_link')
 		));
@@ -302,7 +298,7 @@ class nggManageGallery {
 		$alt_text	= empty($picture->alttext) ? $picture->filename: $picture->alttext;
 		$title		= esc_attr(__('View', 'nggallery'). " \"{$alt_text}\"");
 
-		return "<a href='{$image_url}' class='shutter' title='{$title}'>{$label}</a>";
+		return "<a href='{$image_url}' class='thickbox' title='{$title}'>{$label}</a>";
 	}
 
 	function render_meta_action_link($id, $picture)
@@ -332,41 +328,45 @@ class nggManageGallery {
 		return "<a href='{$url}' class='ngg-dialog' title='{$title}'>{$label}</a>";
 	}
 
-	function render_publish_action_link($id, $picture)
-	{
-		if (!current_user_can('publish_posts')) return FALSE;
-
-		$url		= nextgen_esc_url(NGGALLERY_URLPATH.'admin/publish.php?h=230&id='.$picture->pid);
-		$title		= esc_attr__('Publish this image', 'nggallery');
-		$label		= esc_html__('Publish', 'nggallery');
-
-		return "<a href='{$url}' class='ngg-dialog' title='{$title}'>{$label}</a>";
-	}
-
 	function render_recover_action_link($id, $picture)
 	{
-		if ( !file_exists( $picture->imagePath . '_backup' )) return FALSE;
+		if (!file_exists($picture->imagePath . '_backup'))
+            return FALSE;
 
-		$url		= wp_nonce_url("admin.php?page=nggallery-manage-gallery&amp;mode=recoverpic&amp;gid={$picture->galleryid}&amp;pid={$picture->pid}", 'ngg_recoverpicture');
-		$title		= esc_attr__('Recover image from backup', 'nggallery');
-		$label		= esc_html__('Recover', 'nggallery');
-		$alttext	= empty($picture->alttext) ? $picture->filename : $picture->alttext;
-		$confirm	= addslashes(__("Recover", 'nggallery'). " \"{$alttext}\"?");
-		$onclick	= "javascript:if(!confirm(\"{$confirm}\")) return false";
+		$url      = wp_nonce_url("admin.php?page=nggallery-manage-gallery&amp;mode=recoverpic&amp;gid={$picture->galleryid}&amp;pid={$picture->pid}", 'ngg_recoverpicture');
+		$title    = esc_attr__('Recover image from backup', 'nggallery');
+		$label    = esc_html__('Recover', 'nggallery');
+        $question = __('Recover', 'nggallery');
 
-		return "<a href='{$url}' onclick='{$onclick}' class='confirmrecover' title='{$title}'>{$label}</a>";
+		$alttext = empty($picture->alttext) ? $picture->filename : $picture->alttext;
+		$alttext = M_NextGen_Data::strip_html(html_entity_decode($alttext), TRUE);
+		$alttext = htmlentities($alttext, ENT_QUOTES|ENT_HTML401);
+
+        // Event handler is found in nextgen_admin_page.js
+		return "<a href='{$url}'
+                   class='confirmrecover'
+                   data-question='{$question}'
+                   data-text='{$alttext}'
+                   title='{$title}'>{$label}</a>";
 	}
 
 	function render_delete_action_link($id, $picture)
 	{
-		$url		= wp_nonce_url("admin.php?page=nggallery-manage-gallery&amp;mode=delpic&amp;gid={$picture->galleryid}&amp;pid={$picture->pid}", 'ngg_delpicture');
-		$title		= esc_attr__('Delete image', 'nggallery');
-		$label		= esc_html__('Delete', 'nggallery');
-		$alttext	= empty($picture->alttext) ? $picture->filename : $picture->alttext;
-		$confirm	= addslashes(__("Delete", 'nggallery'). " \"{$alttext}\"?");
-		$onclick	= "javascript:if(!confirm(\"{$confirm}\")) return false;";
+		$url      = wp_nonce_url("admin.php?page=nggallery-manage-gallery&amp;mode=delpic&amp;gid={$picture->galleryid}&amp;pid={$picture->pid}", 'ngg_delpicture');
+		$title    = esc_attr__('Delete image', 'nggallery');
+		$label    = esc_html__('Delete', 'nggallery');
+		$question = __('Delete', 'nggallery');
 
-		return "<a href='{$url}' onclick='{$onclick}' class='submitdelete delete' title='{$title}'>{$label}</a>";
+		$alttext = empty($picture->alttext) ? $picture->filename : $picture->alttext;
+		$alttext = M_NextGen_Data::strip_html(html_entity_decode($alttext), TRUE);
+		$alttext = htmlentities($alttext, ENT_QUOTES|ENT_HTML401);
+
+		// Event handler is found in nextgen_admin_page.js
+		return "<a href='{$url}'
+                   class='submitdelete delete'
+                   data-question='{$question}'
+                   data-text='{$alttext}'
+                   title='{$title}'>{$label}</a>";
 	}
 
 	function render_image_row_header()
@@ -592,6 +592,9 @@ class nggManageGallery {
 								nggGallery::show_message(sprintf(__('One or more "../" in Gallery paths could be unsafe and NextGen Gallery will not delete gallery %s automatically', 'nggallery'), $gallery->{$gallery->id_field}));
 							}
 							else {
+								/**
+								 * @var $mapper Mixin_Gallery_Mapper
+								 */
 								if ($mapper->destroy($id, TRUE))
 									$deleted = TRUE;
 							}
@@ -603,7 +606,6 @@ class nggManageGallery {
 					break;
 			}
 		}
-
 		if (isset ($_POST['addgallery']) && isset ($_POST['galleryname'])){
 
 			check_admin_referer('ngg_addgallery');
@@ -612,12 +614,20 @@ class nggManageGallery {
 				wp_die(__('Cheatin&#8217; uh?', 'nggallery'));
 
 			// get the default path for a new gallery
-			$defaultpath = $ngg->options['gallerypath'];
 			$newgallery = $_POST['galleryname'];
-			if ( !empty($newgallery) )
-				nggAdmin::create_gallery($newgallery, $defaultpath);
+			if (!empty($newgallery))
+			{
+				$gallery_mapper = C_Gallery_Mapper::get_instance();
+				$gallery = $gallery_mapper->create(array('title' => $newgallery));
+				if ($gallery->save() && !isset($_REQUEST['attach_to_post']))
+				{
+					$url = admin_url() . 'admin.php?page=nggallery-manage-gallery&mode=edit&gid=' . $gallery->gid;
+					$message = sprintf(__('Gallery successfully created. <a href="%s" target="_blank">Manage gallery</a>', 'nggallery'), $url);
+					nggGallery::show_message($message, 'gallery_created_msg');
+				}
+			}
 
-            do_action( 'ngg_update_addgallery_page' );
+			do_action( 'ngg_update_addgallery_page' );
 		}
 
 		if (isset ($_POST['TB_bulkaction']) && isset ($_POST['TB_ResizeImages']))  {
@@ -741,10 +751,43 @@ class nggManageGallery {
 
 			switch ($_POST['TB_bulkaction']) {
 				case 'copy_to':
-                    C_Gallery_Storage::get_instance()->copy_images($pic_ids, $dest_gid);
+				    $destination = C_Gallery_Mapper::get_instance()->find($dest_gid);
+                    $new_ids     = C_Gallery_Storage::get_instance()->copy_images($pic_ids, $dest_gid);
+
+                    if (!empty($new_ids))
+                    {
+                        $admin_url = admin_url();
+                        $title     = esc_html($destination->title);
+                        $link      = "<a href='{$admin_url}admin.php?page=nggallery-manage-gallery&mode=edit&gid={$destination->gid}'>{$title}</a>";
+                        nggGallery::show_message(
+                            sprintf(__('Copied %1$s picture(s) to gallery: %2$s .','nggallery'), count($new_ids), $link)
+                        );
+                    }
+                    else {
+                        nggGallery::show_error(
+                            __('Failed to copy images', 'nggallery')
+                        );
+                    }
+
 					break;
 				case 'move_to':
-                    C_Gallery_Storage::get_instance()->move_images($pic_ids, $dest_gid);
+                    $destination = C_Gallery_Mapper::get_instance()->find($dest_gid);
+                    $new_ids     = C_Gallery_Storage::get_instance()->move_images($pic_ids, $dest_gid);
+
+                    if (!empty($new_ids))
+                    {
+                        $admin_url = admin_url();
+                        $title     = esc_html($destination->title);
+                        $link      = "<a href='{$admin_url}admin.php?page=nggallery-manage-gallery&mode=edit&gid={$destination->gid}'>{$title}</a>";
+                        nggGallery::show_message(
+                            sprintf(__('Moved %1$s picture(s) to gallery: %2$s .','nggallery'), count($new_ids), $link)
+                        );
+                    }
+                    else {
+                        nggGallery::show_error(
+                            __('Failed to move images', 'nggallery')
+                        );
+                    }
 					break;
 			}
 		}
@@ -793,12 +836,15 @@ class nggManageGallery {
 			}
 		}
 
-		if (isset ($_POST['updatepictures']) )  {
-		// Update pictures
+		if (isset($_POST['updatepictures']))
+		{
+            // Update pictures
+			$success = FALSE;
 
 			check_admin_referer('ngg_updategallery');
 
-			if ( nggGallery::current_user_can( 'NextGEN Edit gallery options' )  && !isset ($_GET['s']) ) {
+			if (nggGallery::current_user_can('NextGEN Edit gallery options') && !isset($_GET['s']))
+			{
                 $tags = array('<a>', '<abbr>', '<acronym>', '<address>', '<b>', '<base>', '<basefont>', '<big>', '<blockquote>', '<br>', '<br/>', '<caption>', '<center>', '<cite>', '<code>', '<col>', '<colgroup>', '<dd>', '<del>', '<dfn>', '<dir>', '<div>', '<dl>', '<dt>', '<em>', '<fieldset>', '<font>', '<h1>', '<h2>', '<h3>', '<h4>', '<h5>', '<h6>', '<hr>', '<i>', '<img>', '<ins>', '<label>', '<legend>', '<li>', '<menu>', '<noframes>', '<noscript>', '<ol>', '<optgroup>', '<option>', '<p>', '<pre>', '<q>', '<s>', '<samp>', '<select>', '<small>', '<span>', '<strike>', '<strong>', '<sub>', '<sup>', '<table>', '<tbody>', '<td>', '<tfoot>', '<th>', '<thead>', '<tr>', '<tt>', '<u>', '<ul>');
 				$fields = array('title', 'galdesc');
 				
@@ -810,36 +856,51 @@ class nggManageGallery {
 					$html = strip_tags($html, implode('', $tags));
 					$_POST[$field] = $html;
 				}
+
+				$mapper = C_Gallery_Mapper::get_instance();
 				
 				// Update the gallery
-				$mapper = C_Gallery_Mapper::get_instance();
-				if ($entity = $mapper->find($this->gid)) {
-					foreach ($_POST as $key => $value) {
-						$entity->$key = $value;
-					}
-					$mapper->save($entity);
+				if (!$this->gallery)
+				{
+					$this->gallery = $mapper->find($this->gid, TRUE);
 				}
 
-				if ($entity->path == '../' || FALSE !== strpos($entity->path, '/../'))
-					nggGallery::show_message(sprintf(__('One or more "../" in Gallery paths could be unsafe and NextGen Gallery will not delete this gallery automatically', 'nggallery'), $entity->{$entity->id_field}));
+				if ($this->gallery)
+				{
+					foreach ($_POST as $key => $value) {
+						$this->gallery->$key = $value;
+					}
+					$mapper->save($this->gallery);
 
-                wp_cache_delete($this->gid, 'ngg_gallery');
+					if ($this->gallery->is_invalid())
+					{
+						foreach ($this->gallery->get_errors() as $property => $errors) {
+							foreach ($errors as $error) {
+								nggGallery::show_error($error);
+							}
+						}
+					}
 
+					wp_cache_delete($this->gid, 'ngg_gallery');
+					$success = $this->gallery->is_valid();
+				}
 			}
 
-			$this->update_pictures();
-
-			//hook for other plugin to update the fields
-			do_action('ngg_update_gallery', $this->gid, $_POST);
-
-			nggGallery::show_message(__('Updated successfully', 'nggallery'));
+            $pictures_updated = $this->update_pictures();
+			if ($success || $pictures_updated >= 1)
+			{
+				// Hook for other plugin to update the fields
+				do_action('ngg_update_gallery', $this->gid, $_POST);
+				nggGallery::show_message(__('Updated successfully', 'nggallery'));
+			}
 		}
 
-		if (isset ($_POST['scanfolder']))  {
-		// Rescan folder
+        // Rescan folder
+        if (isset ($_POST['scanfolder']))
+        {
 			check_admin_referer('ngg_updategallery');
 
-			$gallerypath = $wpdb->get_var("SELECT path FROM $wpdb->nggallery WHERE gid = '$this->gid' ");
+			$gallerypath = $wpdb->get_var("SELECT `path` FROM {$wpdb->nggallery} WHERE `gid` = '{$this->gid}'");
 			nggAdmin::import_gallery($gallerypath, $this->gid);
 		}
 
@@ -877,51 +938,14 @@ class nggManageGallery {
         }
     }
 
-   	/**
-   	 * Publish a new post with the shortcode from the selected image
-     *
-   	 * @since 1.7.0
-   	 * @return void
-   	 */
-   	function publish_post() {
-
-   	    check_admin_referer('publish-post');
-
-		// Create a WP page
-		global $user_ID, $ngg;
-
-		$ngg->options['publish_width']  = (int) $_POST['width'];
-		$ngg->options['publish_height'] = (int) $_POST['height'];
-		$ngg->options['publish_align'] = $_POST['align'];
-        $align = ( $ngg->options['publish_align'] == 'none') ? '' : 'float='.$ngg->options['publish_align'];
-
-		//save the new values for the next operation
-		update_option('ngg_options', $ngg->options);
-
-		$post['post_type']    = 'post';
-		$post['post_content'] = '[singlepic id=' . intval($_POST['pid']) . ' w=' . $ngg->options['publish_width'] . ' h=' . $ngg->options['publish_height'] . ' ' . $align . ']';
-		$post['post_author']  = $user_ID;
-		$post['post_status']  = isset ( $_POST['publish'] ) ? 'publish' : 'draft';
-		$post['post_title']   = $_POST['post_title'];
-		$post = apply_filters('ngg_add_new_post', $post, $_POST['pid']);
-
-		$post_id = wp_insert_post ($post);
-
-		if ($post_id != 0)
-            nggGallery::show_message( __('Published a new post', 'nggallery') );
-    }
-
 	function can_user_manage_gallery()
 	{
 		$retval 	= FALSE;
-		$registry	= C_Component_Registry::get_instance();
-		$security	= $registry->get_utility('I_Security_Manager');
-		$actor		= $security->get_current_actor();
 
-		if ($this->gallery && $actor->get_entity_id()== $this->gallery->author) {
+		if ($this->gallery && wp_get_current_user()->ID == $this->gallery->author) {
 			$retval = TRUE;
 		}
-		elseif($actor->is_allowed('nextgen_edit_gallery_unowned')) {
+		elseif(M_Security::is_allowed('nextgen_edit_gallery_unowned')) {
 			$retval = TRUE;
 		}
 
@@ -988,7 +1012,9 @@ class nggManageGallery {
             if (!empty($results))
             {
                 foreach ($results as $term_id) {
-                    wp_delete_term($term_id, 'ngg_tag');
+                    $term_id = apply_filters('ngg_pre_delete_unused_term_id', $term_id);
+                    if (!empty($term_id))
+                        wp_delete_term($term_id, 'ngg_tag');
                 }
             }
 		}
@@ -1117,6 +1143,7 @@ class nggManageGallery {
 		$pagination = "<div class='tablenav-pages{$page_class}'>$output</div>";
 
 		echo $pagination;
+		return $pagination;
 	}
 
 }
